Temporarily move RenameFile to org.openrewrite.java
